requirements.txt
add-parent-path
requests
pytest
streamlit
openai
ollama
streamlit_msal
pydantic
streamlit-javascript
sqlalchemy
psycopg2
sqlmodel
loguru
minio
nbformat
nbconvert
python-dotenv
PyYAML
# chromadb-client
# langchain-chroma
# sentence-transformers
# langchain-huggingface
# langchain-text-splitters
# huggingface_hub
